[SDSAET-38] Move W3chtml5Error models to models-api module
