+++ api/models-api/src/main/java/com/cognifide/aet/models/w3chtml5/W3cHtml5ComparatorResult.java
@@ -1,4 +1,4 @@
-/**
+/*
  * AET
  *
  * Copyright (C) 2013 Cognifide Limited
@@ -13,8 +13,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.job.common.comparators.w3chtml5;
-
+package com.cognifide.aet.models.w3chtml5;
 
 import java.io.Serializable;
 import java.util.List;
@@ -31,7 +30,10 @@ public class W3cHtml5ComparatorResult implements Serializable {
 
   private final List<W3cHtml5Issue> excludedIssues;
 
-  public W3cHtml5ComparatorResult(int errorCount, int warningCount, List<W3cHtml5Issue> issues,
+  public W3cHtml5ComparatorResult(
+      int errorCount,
+      int warningCount,
+      List<W3cHtml5Issue> issues,
       List<W3cHtml5Issue> excludedIssues) {
     this.errorCount = errorCount;
     this.warningCount = warningCount;
