+++ core/jobs/src/main/java/com/cognifide/aet/job/common/comparators/w3chtml5/parser/W3cHtml5IssueDeserializer.java
@@ -15,8 +15,8 @@
  */
 package com.cognifide.aet.job.common.comparators.w3chtml5.parser;
 
-import com.cognifide.aet.job.common.comparators.w3chtml5.W3cHtml5Issue;
-import com.cognifide.aet.job.common.comparators.w3chtml5.W3cHtml5IssueType;
+import com.cognifide.aet.models.w3chtml5.W3cHtml5Issue;
+import com.cognifide.aet.models.w3chtml5.W3cHtml5IssueType;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
