+++ core/jobs/src/main/java/com/cognifide/aet/job/common/datafilters/w3chtmlfilter/W3cHtml5IssuesFilter.java
@@ -18,11 +18,11 @@ package com.cognifide.aet.job.common.datafilters.w3chtmlfilter;
 import com.cognifide.aet.job.api.datafilter.AbstractDataModifierJob;
 import com.cognifide.aet.job.api.exceptions.ParametersException;
 import com.cognifide.aet.job.api.exceptions.ProcessingException;
-import com.cognifide.aet.job.common.comparators.w3chtml5.W3cHtml5ComparatorResult;
-import com.cognifide.aet.job.common.comparators.w3chtml5.W3cHtml5Issue;
 import com.cognifide.aet.job.common.comparators.w3chtml5.W3cHtml5IssueComparator;
-import com.cognifide.aet.job.common.comparators.w3chtml5.W3cHtml5IssueType;
 import com.cognifide.aet.job.common.utils.ParamsHelper;
+import com.cognifide.aet.models.w3chtml5.W3cHtml5ComparatorResult;
+import com.cognifide.aet.models.w3chtml5.W3cHtml5Issue;
+import com.cognifide.aet.models.w3chtml5.W3cHtml5IssueType;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
